+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.0.2
 
-- Fix error when reconnecing the client
+- Fix error when reconnecting the client
 - Fix last modified timestamps using the remote node id
+- Add optional WebSocket heartbeat (on by default)
 
 ## 0.0.1+1